requirements.txt
pyyaml
numpy==1.26.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==1.12.1+cpu
torchmetrics==0.9.3
matplotlib
scikit-learn==1.5.1
ipykernel
pandas
tqdm
